feat: split settingscontent into multiple smaller components
